AddressController changes
- Update and destroy action are available
- In ProductController, we verify first that the product exists before
destroying it
- AddressRequest modified to return errors in JSON

+++ app/Http/Controllers/ProductController.php
@@ -88,6 +88,9 @@ class ProductController extends Controller
     public function destroy($id)
     {
         $product = App\Product::find($id);
-        $product->delete();
+
+        if ($product){
+            $product->delete();
+        }
     }
 }
